Separated for maintainability
